+++ src/test/java/juuxel/vineflowerforloom/test/oldapi/BasicProjectTest.java
@@ -1,5 +1,6 @@
-package juuxel.loomquiltflower.test;
+package juuxel.vineflowerforloom.test.oldapi;
 
+import juuxel.vineflowerforloom.test.ProjectTest;
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.GradleRunner;
 import org.gradle.testkit.runner.TaskOutcome;
@@ -7,11 +8,11 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class SnapshotFromRepositoriesTest extends ProjectTest {
+class BasicProjectTest extends ProjectTest {
     @Test
     void test() {
         // Set up
-        setupProject("snapshot-from-repositories");
+        setupProject("deprecated/basic");
 
         // Run
         BuildResult result = GradleRunner.create()
@@ -22,7 +23,6 @@ class SnapshotFromRepositoriesTest extends ProjectTest {
             .withDebug(true)
             .build();
 
-        assertThat(result.getOutput()).contains("Resolved: 1.8.1-20220429.000346-3");
         assertThat(result.task(":genSourcesWithQuiltflower").getOutcome()).isEqualTo(TaskOutcome.SUCCESS);
     }
 }
