+++ src/test/java/juuxel/vineflowerforloom/test/oldapi/RuntimeClasspathTest.java
@@ -1,5 +1,6 @@
-package juuxel.loomquiltflower.test;
+package juuxel.vineflowerforloom.test.oldapi;
 
+import juuxel.vineflowerforloom.test.ProjectTest;
 import org.gradle.testkit.runner.BuildResult;
 import org.gradle.testkit.runner.GradleRunner;
 import org.gradle.testkit.runner.TaskOutcome;
@@ -7,22 +8,22 @@ import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-class SnapshotFromQuiltTest extends ProjectTest {
+class RuntimeClasspathTest extends ProjectTest {
     @Test
     void test() {
         // Set up
-        setupProject("snapshot-from-quilt");
+        setupProject("deprecated/runtime-classpath");
 
         // Run
         BuildResult result = GradleRunner.create()
             .withPluginClasspath()
             .withProjectDir(projectDirectory)
-            .withArguments("genSourcesWithQuiltflower", "--stacktrace")
+            .withArguments("verifyRuntimeClasspath", "genSourcesWithQuiltflower", "--stacktrace")
             .forwardOutput()
             .withDebug(true)
             .build();
 
-        assertThat(result.getOutput()).contains("Resolved: 1.9.0-20221030.213144-213");
         assertThat(result.task(":genSourcesWithQuiltflower").getOutcome()).isEqualTo(TaskOutcome.SUCCESS);
+        assertThat(result.task(":verifyRuntimeClasspath").getOutcome()).isEqualTo(TaskOutcome.SUCCESS);
     }
 }
