+++ src/test/resources/projects/deprecated/basic/settings.gradle
@@ -1,3 +1,5 @@
+package projects.deprecated.basic
+
 pluginManagement {
     repositories {
         maven {
