+++ src/test/resources/projects/deprecated/preferences/gradle.properties
