+++ src/test/resources/projects/deprecated/preferences/settings.gradle
@@ -1,3 +1,5 @@
+package projects.deprecated.preferences
+
 pluginManagement {
     repositories {
         maven {
