+++ src/test/resources/projects/deprecated/runtime-classpath/build.gradle
