+++ src/test/resources/projects/deprecated/runtime-classpath/settings.gradle
