+++ src/test/resources/projects/deprecated/snapshot-from-quilt/build.gradle
