+++ src/test/resources/projects/deprecated/snapshot-from-quilt/settings.gradle
