+++ src/test/resources/projects/deprecated/snapshot-from-repositories/build.gradle
