+++ src/test/resources/projects/deprecated/snapshot-from-repositories/settings.gradle
